+++ crates/table/src/read_column.rs
@@ -328,7 +328,7 @@ macro_rules! impl_read_column_via_from {
 }
 
 impl_read_column_via_from! {
-    u32 => spacetimedb_primitives::ColId;
+    u16 => spacetimedb_primitives::ColId;
     u32 => spacetimedb_primitives::TableId;
     u32 => spacetimedb_primitives::IndexId;
     u32 => spacetimedb_primitives::ConstraintId;
